+++ src/main/java/org/seedstack/feign/FeignConfig.java
@@ -5,17 +5,8 @@
  * License, v. 2.0. If a copy of the MPL was not distributed with this
  * file, You can obtain one at http://mozilla.org/MPL/2.0/.
  */
+
 package org.seedstack.feign;
-
-import java.net.URL;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
-import javax.validation.constraints.NotNull;
-
-import org.seedstack.coffig.Config;
-import org.seedstack.coffig.SingleValue;
 
 import feign.Contract;
 import feign.Logger;
@@ -26,9 +17,21 @@ import feign.codec.Encoder;
 import feign.jackson.JacksonDecoder;
 import feign.jackson.JacksonEncoder;
 import feign.slf4j.Slf4jLogger;
+import java.net.URL;
+import java.util.Collections;
+import java.util.HashMap;
+import java.util.Map;
+import java.util.concurrent.TimeUnit;
+import javax.validation.constraints.Min;
+import javax.validation.constraints.NotNull;
+import org.seedstack.coffig.Config;
+import org.seedstack.coffig.SingleValue;
+import org.seedstack.feign.internal.FeignErrorCode;
+import org.seedstack.seed.SeedException;
 
 @Config("feign")
 public class FeignConfig {
+    @NotNull
     private Map<Class<?>, EndpointConfig> endpoints = new HashMap<>();
 
     public Map<Class<?>, EndpointConfig> getEndpoints() {
@@ -44,21 +47,25 @@ public class FeignConfig {
         @SingleValue
         @NotNull
         private URL baseUrl;
-
         private Class<? extends Contract> contract;
-
+        @NotNull
         private Class<? extends Target> target = HardCodedTarget.class;
-
+        @NotNull
         private Class<? extends Encoder> encoder = JacksonEncoder.class;
-
+        @NotNull
         private Class<? extends Decoder> decoder = JacksonDecoder.class;
-
+        @NotNull
         private Class<? extends Logger> logger = Slf4jLogger.class;
-
+        @NotNull
         private Logger.Level logLevel = Logger.Level.NONE;
-
+        @NotNull
+        private TimeUnit timeUnit = TimeUnit.MILLISECONDS;
+        @Min(0)
+        private int connectTimeout = 10000;
+        @Min(0)
+        private int readTimeout = 60000;
+        @NotNull
         private HystrixWrapperMode hystrixWrapper = HystrixWrapperMode.AUTO;
-
         private Class<?> fallback;
 
         public URL getBaseUrl() {
@@ -92,10 +99,22 @@ public class FeignConfig {
             return decoder;
         }
 
-        public Class<? extends Target> getTarget() {
-            return target;
+        @SuppressWarnings("unchecked")
+        public <T> Class<? extends Target<T>> getTarget(Class<T> apiClass) {
+            if (target == null) {
+                return null;
+            } else {
+                if (HardCodedTarget.class.isAssignableFrom(target) || apiClass.isAssignableFrom(target)) {
+                    return (Class<Target<T>>) target;
+                } else {
+                    throw SeedException
+                            .createNew(FeignErrorCode.BAD_TARGET_CLASS)
+                            .put("class", target)
+                            .put("api", apiClass);
+                }
+            }
         }
-        
+
         public EndpointConfig setTarget(Class<? extends Target> target) {
             this.target = target;
             return this;
@@ -133,12 +152,40 @@ public class FeignConfig {
             return this;
         }
 
-        public Class<?> getFallback() {
-            return fallback;
+        @SuppressWarnings("unchecked")
+        public <T> Class<T> getFallback() {
+            return (Class<T>) fallback;
         }
 
         public EndpointConfig setFallback(Class<?> fallback) {
             this.fallback = fallback;
+            return this;
+        }
+
+        public TimeUnit getTimeUnit() {
+            return timeUnit;
+        }
+
+        public EndpointConfig setTimeUnit(TimeUnit timeUnit) {
+            this.timeUnit = timeUnit;
+            return this;
+        }
+
+        public int getConnectTimeout() {
+            return connectTimeout;
+        }
+
+        public EndpointConfig setConnectTimeout(int connectTimeout) {
+            this.connectTimeout = connectTimeout;
+            return this;
+        }
+
+        public int getReadTimeout() {
+            return readTimeout;
+        }
+
+        public EndpointConfig setReadTimeout(int readTimeout) {
+            this.readTimeout = readTimeout;
             return this;
         }
     }