+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+# Version 1.1.0 (2017-11-31)
+
+* [new] Support fallback classes based on `FallbackFactory` to access the exception if any.
+* [new] Add configuration options for timeouts. 
+
 # Version 1.0.0 (2017-05-02)
 
 * [new] Extracted from the Netflix add-on.